Refactor error handling

+++ controllers/appointmentController.js
@@ -2,6 +2,11 @@ const emailService = require("../services/emailService");
 const bookingService = require("../services/bookingService");
 const slotService = require("../services/slotService");
 const appointmentService = require("../services/appointmentService");
+const {
+  EmailServiceError,
+  QueryError,
+  MessageCreationError,
+} = require("../services/errorService");
 
 const createAppointment = async (req, res) => {
   const {
@@ -15,6 +20,20 @@ const createAppointment = async (req, res) => {
   } = req.body;
 
   try {
+    if (
+      !firstName ||
+      !lastName ||
+      !email ||
+      !selectedDate ||
+      !selectedSlot ||
+      !services ||
+      !prestations
+    ) {
+      return res
+        .status(400)
+        .json({ message: "Tous les champs sont obligatoires." });
+    }
+
     const { userMessage, adminMessage, appointmentId } =
       await bookingService.createBooking(
         firstName,
@@ -25,7 +44,35 @@ const createAppointment = async (req, res) => {
         prestations
       );
 
-    await appointmentService.insertAppointmentServices(appointmentId, services);
+    try {
+      await appointmentService.insertAppointmentServices(
+        appointmentId,
+        services
+      );
+      await slotService.updateSlotWithAppointmentId(
+        selectedSlot,
+        appointmentId
+      );
+      await slotService.updateSlotToBooked(selectedSlot);
+    } catch (error) {
+      if (error instanceof QueryError) {
+        console.error(
+          "Erreur lors de la mise à jour du rendez-vous ou des créneaux : ",
+          error
+        );
+        return res.status(500).json({
+          message: error.message,
+        });
+      }
+      console.error(
+        "Erreur lors de la mise à jour du rendez-vous ou des créneaux : ",
+        error
+      );
+      return res.status(500).json({
+        message:
+          "Erreur lors de la mise à jour du rendez-vous ou des créneaux.",
+      });
+    }
 
     await emailService.sendEmail(
       email,
@@ -39,12 +86,16 @@ const createAppointment = async (req, res) => {
       adminMessage
     );
 
-    await slotService.updateSlotWithAppointmentId(selectedSlot, appointmentId);
-
-    await slotService.updateSlotToBooked(selectedSlot);
-
     res.status(200).json({ message: "Rendez-vous créé et emails envoyés." });
   } catch (error) {
+    if (
+      error instanceof EmailServiceError ||
+      error instanceof SlotError ||
+      error instanceof MessageCreationError ||
+      error instanceof QueryError
+    ) {
+      return res.status(500).json({ message: error.message });
+    }
     console.error(error);
     res
       .status(500)
@@ -57,8 +108,15 @@ const getAppointments = async (req, res) => {
     const appointments = await appointmentService.getAppointments();
     res.json(appointments);
   } catch (error) {
+    if (error instanceof QueryError) {
+      console.error("Erreur lors de la récupération des rendez-vous : ", error);
+      return res.status(500).json({
+        message: error.message,
+      });
+    }
+    console.error("Erreur lors de la récupération des rendez-vous : ", error);
     res.status(500).json({
-      error: "Erreur lors de la récupération des rendez-vous",
+      message: "Erreur lors de la récupération des rendez-vous.",
     });
   }
 };
@@ -68,8 +126,21 @@ const getAppointmentsWithDetails = async (req, res) => {
     const appointments = await appointmentService.getAppointmentsWithDetails();
     res.json(appointments);
   } catch (error) {
+    if (error instanceof QueryError) {
+      console.error(
+        "Erreur lors de la récupération des rendez-vous détaillés : ",
+        error
+      );
+      return res.status(500).json({
+        message: error.message,
+      });
+    }
+    console.error(
+      "Erreur lors de la récupération des rendez-vous détaillés : ",
+      error
+    );
     res.status(500).json({
-      error: "Erreur lors de la récupération des rendez-vous",
+      message: "Erreur lors de la récupération des rendez-vous détaillés.",
     });
   }
 };
@@ -80,6 +151,14 @@ const confirmAppointment = async (req, res) => {
     await appointmentService.confirmAppointment(appointmentId, appointment);
     res.status(200).json({ message: "Rendez-vous confirmé et email envoyé" });
   } catch (error) {
+    console.error("Erreur lors de la confirmation du rendez-vous : ", error);
+    if (
+      error instanceof QueryError ||
+      error instanceof EmailServiceError ||
+      error instanceof MessageCreationError
+    ) {
+      return res.status(500).json({ message: error.message });
+    }
     res
       .status(500)
       .json({ message: "Erreur lors de la confirmation du rendez-vous" });
@@ -91,34 +170,86 @@ const getgroupedAppointments = async (req, res) => {
     const appointments = await appointmentService.getgroupedAppointments();
     res.json(appointments);
   } catch (error) {
+    console.error(
+      "Erreur lors de la récupération des rendez-vous groupés : ",
+      error
+    );
+    if (error instanceof QueryError) {
+      return res.status(500).json({
+        message: error.message,
+      });
+    }
     res.status(500).json({
-      error: "Erreur lors de la récupération des rendez-vous",
+      message: "Erreur lors de la récupération des rendez-vous groupés",
     });
   }
 };
 
 const deleteAppointmentServices = async (req, res) => {
+  const appointmentId = req.params.appointmentId;
+
   try {
-    const appointmentId = req.params.appointmentId;
+    const appointment = await appointmentService.getAppointmentById(
+      appointmentId
+    );
+
+    if (!appointment) {
+      return res.status(404).json({ message: "Rendez-vous non trouvé." });
+    }
+
     await appointmentService.deleteAppointmentServices(appointmentId);
+
     res.status(200).json({
       message: "Les services du rendez-vous ont été supprimés avec succès.",
     });
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    console.error(
+      "Erreur lors de la suppression des prestations du rendez-vous :",
+      error
+    );
+    if (error instanceof QueryError) {
+      return res.status(500).json({
+        message: error.message,
+      });
+    }
+    res.status(500).json({
+      message: "Erreur lors de la suppression des prestations du rendez-vous.",
+    });
   }
 };
 
 const addAppointmentServices = async (req, res) => {
+  const appointmentId = req.params.appointmentId;
+  const services = req.body.services;
+
   try {
-    const appointmentId = req.params.appointmentId;
-    const services = req.body.services;
+    const appointment = await appointmentService.getAppointmentById(
+      appointmentId
+    );
+    if (!appointment) {
+      return res.status(404).json({ message: "Rendez-vous non trouvé." });
+    }
+
     await appointmentService.addAppointmentServices(appointmentId, services);
+
     res.status(200).json({
-      message: "Les services ont été ajoutés au rendez-vous avec succès.",
+      message: "Les prestations ont été ajoutés au rendez-vous avec succès.",
     });
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    console.error(
+      "Erreur lors de l'ajout des prestations au rendez-vous :",
+      error
+    );
+    if (error instanceof QueryError) {
+      return res.status(500).json({
+        message: error.message,
+      });
+    }
+    res
+      .status(500)
+      .json({
+        message: "Erreur lors de l'ajout des prestations au rendez-vous.",
+      });
   }
 };
